pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "abbibench_toolkit"
version = "2025.09.12"
description = "abbibench made by Jaejoon Kim"
readme = "README.md"
requires-python = "==3.10.14"
dependencies = [
	"torch==2.2.0",
	"torchvision==0.17.0",
	"torchaudio==2.2.0",
	"static_error_handler",
	"tyro",
	"tqdm",
	"black",
	"mypy",
	"numpy==1.26.4",
	"biopython==1.80",
	"biotite==0.41.2",
	"transformers==4.48.1",
	"pandas",
	"scipy",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index="pytorch-cu118" },
  ]

torchvision = [
  { index = "pytorch-cu118" },
  ]

torchaudio = [
{ index = "pytorch-cu118" },
]

[tool.hatch.build.targets.wheel]
packages = ["src/abbibench_toolkit"]

[tool.hatch.metadata]
allow-direct-references = true
